Split out the Dimension trait
